Update README.md
So these links are broken, which I assume has to do with the "default" branch being changed to "testnet3". For now, this will solve the problem, but we probably want to create a more scalable way to do this or set up a process to check that these are updated when/if the default branch changes.

For /wasm/pkg I just put the same URL as wasm, since the actual npm package is linked.

Signed-off-by: a h <austin@aleo.org>