Update SnapshotQuery to use new batch statepaths fetching method + upgrade to SnarkVM v4.2.0
With SnarkOS version 4.1.0 a new batch statepaths endpoint was added, this PR adds the SnapShot query to fetch directly from this endpoint due to the SnarkVM query's insufficient design.

Further the upgrade to SnarkVM v4.2.0 requires addition of consensus versions as a method argument to verification routines and deployment cost calculations.
To account for this, this PR adds automatic detection of ConsensusVersion and InclusionVersion based on block height detected from the Query.