Fix import resolution and proof generation when building from Authorizations.
Previously when building from authorizations, the default SnarkVM Query did not work as expected in the wasm environment.

This is partially due to errors in URL parsing in the Query object and incompatibilities with polyfills the SDK uses. This
PR adds compatible REST functionality to the Snapshot Query to allow it to build queries properly. It also fixes import
resolution by using the add_program_with_edition function.

This PR also adds convenience methods for estimating fees to the programManager and provides example documentation for using them.
Documentation within the ProgramManager for using these functions and proper usage of the new buildTransactionFromAuthorization
and adds a new create-leo-app example that shows how to use these functions properly.